Prevent null user agent being passed to DeviceDetector
`request()->userAgent()` may return a null value if the user agent is not available in the HTTP request. This can happen in some situations, such as when using a console application or when requesting through an API.

The null coalescing operator (??) ensures that arguments passed to the DeviceDetector constructor are never null. If `request()->userAgent()` returns null, this operator will return an empty string ('') instead.