fix for usernames that don't appear in the table
To display the username, we previously relied on the route name obtained from the request. But apparently, the table did not make the corresponding request, so when we moved the page, the username could not be displayed.

Now we rely on the table name to display the username only in the all urls table.