fix: Initialize user_id property in UrlTableByUser
The typed property `$user_id` in the `UrlTableByUser` Livewire component was accessed in the `scopeDatasource` method before it was guaranteed to be initialized, leading to a "Typed property must not be accessed before initialization" error.

This commit initializes `$user_id` to `null` by default upon declaration to prevent this error.

Related:
https://github.com/realodix/urlhub/pull/1116